[DEV-1170] added next collateral ratio, calculations and moved everything to a new service
